% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GNSSseg.R
\name{GNSSseg}
\alias{GNSSseg}
\title{Homogeneization of GNSS series}
\usage{
GNSSseg(Data, lyear = 365.25, lmin = 1, Kmax = NULL,
  selection.K = "mBIC", S = 0.75, f = TRUE, selection.f = FALSE,
  threshold = 0.05, tol = 1e-04)
}
\arguments{
\item{Data}{a data frame, with size [n x 2], containing the signal (e.g. the daily GPS-ERAI series for GNSS) and the dates (in format yyyy-mm-dd of type "calendar time" (class POSIXct))}

\item{lyear}{the length of the year in the signal. Defalut is 365.25}

\item{lmin}{the minimum length of the segments. Default is 1}

\item{Kmax}{the maximal number of segments (must be lower than n). Default is [n/2 log(n)]}

\item{selection.K}{a name indicating the model selection criterion to select the number of segments K (\code{mBIC}, \code{ML}, \code{BM_BJ} or \code{BM_slope}). \code{"none"} indicates that no selection is claimed and the procedure considers \code{Kmax} segments or \code{Kmax}-1 changes. If \code{selection.K="All"}, the results for the four possible criteria are given. Default is \code{"mBIC"}}

\item{S}{the threshold used in the ML criterion. Default is 0.75}

\item{f}{a boolean indicating if the functional part is taking into account for in the model. Default is FALSE}

\item{selection.f}{a boolean indicating if a selection on the functions of the Fourier decomposition of order 4 is performed. Default is FALSE}

\item{threshold}{a numeric value lower than 1. Default is 0.05}

\item{tol}{the stopping rule for the iterative procedure. Default is 1e-4}
}
\value{
A file containing
\itemize{
\item \code{selected.K} that corresponds to the selected number of segments. If \code{selection.K="none"}, the number of segments is fixed to \code{Kmax}
\item \code{segmentation} that corresponds to the estimation of the segmentation parameters (the begin and the end positions of each segment with the estimated mean)
\item \code{functional} that corresponds to the estimation of the functional part
\item \code{loglik} that corresponds to the log-likelihood for k=1,...,\code{Kmax}. If \code{selection.K="none"}, it contains only the log-likelihood for \code{Kmax} segments
\item \code{variances} that corresponds to the estimated variances for each fixed interval (e.g. the months).
\item \code{mBIC} that corresponds to the values of the mBIC criterion for k=1,...,\code{Kmax} if it is required (\code{selection.K="mBIC"} or \code{selection.K="All"})
}
If \code{selection.K="All"}, the outputs \code{selected.K}, \code{segmentation} and \code{functional} are each a list containing the corresponding result for the four model selection criteria
}
\description{
fit a segmentation in the mean model by taken into account for a functional part (estimated with a Fourier decomposition) and a heterogeneous variance (monthly variance)
}
\details{
The function performs homogeneization of GNSS series. The considered model is a segmentation in the mean model (detection of abrupt changes model) in which a functional part is added and with heterogeneous variances on fixed intervals. On GNSS series, the latter intervals corresponds to the month.
The inference procedure consists for a fixed number of segments K in first estimating robustly the variances and second estimating the segmentation and the functional parts using an iterative procedure. The solution is obtained for k=1,...,\code{Kmax} segments. Then the "best k" is chosen using model selection criteria. The possible criteria are \code{mBIC} the modified BIC criterion REFEREF, \code{ML} the criterion proposed by REFEF, \code{BM_BJ} and \code{BM_slope} the criteria proposed by REFEF where the penalty constant is calibrated using the Biggest Jump and the slope respectively REFERF.
\itemize{
\item The data is a data frame with 2 columns: $signal is the signal to be homogeneized (a daily series) and $date is the date. The date will be in format yyyy-mm-dd of type "calendar time" (class POSIXct).
\item The function part is estimated using a Fourier decomposition of order 4 with \code{selection.f=FALSE}. \code{selection.f=TRUE} consists in selecting twice functions of the Fourier decomposition of order 4, first by using a stepwise procedure using AIC and second by considering among them the signficative ones (for which p.values are lower than \code{threshold})
\item If \code{selection.K="none"} the procedure is performed with \code{Kmax} segments.
\item Missing data in the signal are accepted.
}
Note that if \code{f=FALSE}, only a segmentation is performed.
}
\examples{
data(Data)
lyear=365.25
Kmax=8
lmin=1
result=GNSSseg(Data,lyear,Kmax=Kmax)
plot_GNSS(Data,result$segmentation,result$functional)
}
